Refactor pattern beautifiers.
