+++ opencolorio_config_aces/config/generation/__init__.py
@@ -1,6 +1,24 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
+from .beautifiers import (
+    SEPARATOR_COLORSPACE_NAME,
+    SEPARATOR_COLORSPACE_FAMILY,
+    SEPARATOR_BUILTIN_TRANSFORM_NAME,
+    PATTERNS_COLORSPACE_NAME,
+    PATTERNS_LOOK_NAME,
+    PATTERNS_TRANSFORM_FAMILY,
+    PATTERNS_VIEW_TRANSFORM_NAME,
+    PATTERNS_DISPLAY_NAME,
+    PATTERNS_ALIAS,
+    beautify_name,
+    beautify_colorspace_name,
+    beautify_look_name,
+    beautify_transform_family,
+    beautify_view_transform_name,
+    beautify_display_name,
+    beautify_alias,
+)
 from .factories import (
     group_transform_factory,
     colorspace_factory,
@@ -21,6 +39,24 @@ from .common import (
 )
 
 __all__ = [
+    "SEPARATOR_COLORSPACE_NAME",
+    "SEPARATOR_COLORSPACE_FAMILY",
+    "SEPARATOR_BUILTIN_TRANSFORM_NAME",
+    "PATTERNS_COLORSPACE_NAME",
+    "PATTERNS_LOOK_NAME",
+    "PATTERNS_TRANSFORM_FAMILY",
+    "PATTERNS_VIEW_TRANSFORM_NAME",
+    "PATTERNS_DISPLAY_NAME",
+    "PATTERNS_ALIAS",
+    "beautify_name",
+    "beautify_colorspace_name",
+    "beautify_look_name",
+    "beautify_transform_family",
+    "beautify_view_transform_name",
+    "beautify_display_name",
+    "beautify_alias",
+]
+__all__ += [
     "group_transform_factory",
     "colorspace_factory",
     "named_transform_factory",
